adding comments
So essentially this Python script reads in our CSV file and outputs a JSON file. It needs a little tinkering because I don't know what format we need the JSON file in, but that should be easily worked out. 
I'm still working out what to do with this JSON file once it's been generated. Is it interacted upon by JQuery? Not even sure how that works haha.